+++ bindings/nostr-sdk-ffi/src/protocol/event/tag/standard.rs
@@ -37,6 +37,8 @@ pub enum TagStandard {
         marker: Option<Marker>,
         /// Should be the public key of the author of the referenced event
         public_key: Option<Arc<PublicKey>>,
+        /// Whether the e tag is an uppercase E or not
+        uppercase: bool,
     },
     /// Git clone (`clone` tag)
     ///
@@ -108,6 +110,8 @@ pub enum TagStandard {
     },
     Kind {
         kind: KindEnum,
+        /// Whether the k tag is an uppercase K or not
+        uppercase: bool,
     },
     RelayUrl {
         relay_url: String,
@@ -279,11 +283,13 @@ impl From<tag::TagStandard> for TagStandard {
                 relay_url,
                 marker,
                 public_key,
+                uppercase,
             } => Self::EventTag {
                 event_id: Arc::new(event_id.into()),
                 relay_url: relay_url.map(|u| u.to_string()),
                 marker: marker.map(|m| m.into()),
                 public_key: public_key.map(|p| Arc::new(p.into())),
+                uppercase,
             },
             tag::TagStandard::GitClone(urls) => Self::GitClone {
                 urls: urls.into_iter().map(|r| r.to_string()).collect(),
@@ -351,7 +357,10 @@ impl From<tag::TagStandard> for TagStandard {
             tag::TagStandard::ExternalIdentity(identity) => Self::ExternalIdentity {
                 identity: identity.into(),
             },
-            tag::TagStandard::Kind(kind) => Self::Kind { kind: kind.into() },
+            tag::TagStandard::Kind { kind, uppercase } => Self::Kind {
+                kind: kind.into(),
+                uppercase,
+            },
             tag::TagStandard::Relay(url) => Self::RelayUrl {
                 relay_url: url.to_string(),
             },
@@ -475,11 +484,13 @@ impl TryFrom<TagStandard> for tag::TagStandard {
                 relay_url,
                 marker,
                 public_key,
+                uppercase,
             } => Ok(Self::Event {
                 event_id: **event_id,
                 relay_url: relay_url.map(UncheckedUrl::from),
                 marker: marker.map(nip10::Marker::from),
                 public_key: public_key.map(|p| **p),
+                uppercase,
             }),
             TagStandard::GitClone { urls } => {
                 let mut parsed_urls: Vec<Url> = Vec::with_capacity(urls.len());
@@ -544,7 +555,10 @@ impl TryFrom<TagStandard> for tag::TagStandard {
                 coordinate: coordinate.as_ref().deref().clone(),
                 relay_url: relay_url.map(UncheckedUrl::from),
             }),
-            TagStandard::Kind { kind } => Ok(Self::Kind(kind.into())),
+            TagStandard::Kind { kind, uppercase } => Ok(Self::Kind {
+                kind: kind.into(),
+                uppercase,
+            }),
             TagStandard::RelayUrl { relay_url } => Ok(Self::Relay(UncheckedUrl::from(relay_url))),
             TagStandard::POW { nonce, difficulty } => Ok(Self::POW {
                 nonce: nonce.parse()?,