+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 
 * nostr: add `SingleLetterTag::as_str` and `TagKind::as_str` ([Yuki Kishimoto])
 * nostr: add `Kind::Comment` ([reyamir])
+* nostr: add `EventBuilder::comment` ([reyamir])
 
 ### Fixed